EST-1017: moves InvoiceDtoFactory and InvoiceItemDtoFactory under org.estatio.dom
