+++ estatioapp/dom/src/test/java/org/estatio/dom/invoice/canonical/v1/InvoiceItemDtoFactory_Test.java
@@ -1,4 +1,4 @@
-package org.estatio.canonical.invoice;
+package org.estatio.dom.invoice.canonical.v1;
 
 import org.joda.time.LocalDate;
 import org.junit.Before;
@@ -10,6 +10,7 @@ import org.estatio.dom.asset.Unit;
 import org.estatio.dom.charge.Charge;
 import org.estatio.dom.charge.ChargeGroup;
 import org.estatio.dom.invoice.Invoice;
+import org.estatio.dom.invoice.canonical.v1.InvoiceItemDtoFactory;
 import org.estatio.dom.lease.Lease;
 import org.estatio.dom.lease.Occupancy;
 import org.estatio.dom.lease.invoicing.InvoiceItemForLease;
